+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.105.0] - Not released
+### Fixed
+- A group administrator could not delete a message with an empty body (and with
+  attachments) from a managed group.
+
 ## [1.104.2] - 2021-12-13
 ### Fixed
 - Allow to edit direct message without recipients.